chore(layout): move mtp_forward to compute/, vision_pipeline to vision/
Phase 4 soft refactor. Two file moves that align files with their
actual responsibility:

- src/runtime/mtp_forward.{cu,h} → src/compute/mtp_forward.{cu,h}
  (kernel file, not orchestration — belongs with the other compute kernels)

- src/runtime/vision_pipeline.{cpp,h} → src/vision/vision_pipeline.{cpp,h}
  (aligns with the other vision/ files: vision_encoder, vision_loader,
  image_processor, vision_model)

5 include sites updated (engine.h, engine.cpp, engine_scheduler.cpp,
vision_pipeline.cpp, mtp_forward.cu, tests/test_mtp_forward.cpp).
CMakeLists.txt sources moved.

make build green.

Phase 4 soft PRs (deferred) of docs/superpowers/specs/2026-05-20-architecture-refactor-roadmap-design.md
